requirements.txt
pymorphy2
regex
nltk
pony
vk-api
flask
GitPython